requirements.txt
fastapi==0.109.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
uvicorn==0.27.0.post1
pymongo==4.6.1